+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 ### Added
 - Added 4.0 Version of Android SDK which has Identify API features IdentifyAddress, IdentifyRisk, IdentifyEmail and IdentifyExtract.
 - Added 4.1 Version of Android SDK. Added GetCityStateProvinceAPI and GetPostalCodesAPI under IdentifyAddress.
+- Added 4.2 Version of Android SDK. Added ValidateMailingAddressUSCANAPI under IdentifyAddress.